+++ CHANGELOG.md
@@ -43,6 +43,7 @@ from the default profiles: `shift-heading-level-by: 1`.
 - Zettlr now remembers the widths of file manager and sidebar
 - You can now reset the file manager and sidebar widths by double-clicking the
   corresponding resizer
+- Dark mode folded header's ellipsis are now properly colored.
 - Copying plain links in the form `<http://www.example.com>` will now remove the
   angled brackets (#5285)
 - Reverted a change from 3.1.0 which altered the process of creating new files